requirements.txt
fastapi
uvicorn[standard]
python-multipart
pywebpush
cryptography
python-dotenv
